WIP: building infra to analyze profit on a spread option trade

+++ src/eventmodels/tradier_order_update_event.go
@@ -1,7 +1,7 @@
 package eventmodels
 
 type TradierOrderUpdateEvent struct {
-	OrderID uint64
+	OrderID uint
 	Field   string
 	Old     interface{}
 	New     interface{}
